+++ ironfish-cli/src/commands/multisig/account/identities.ts
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { Flags } from '@oclif/core'
-import { IronfishCommand } from '../../../../command'
-import { RemoteFlags } from '../../../../flags'
+import { IronfishCommand } from '../../../command'
+import { RemoteFlags } from '../../../flags'
 
 export class MultisigAccountIdentities extends IronfishCommand {
   static description = `List the identities for a multisig account`
